Include SNS tokens and ICP
